+++ react/components/MinicartIconButton.tsx
@@ -11,22 +11,24 @@ const CSS_HANDLES = ['minicartIconContainer', 'minicartQuantityBadge'] as const
 
 interface MinicartIconButtonProps {
   quantityDisplay: MinicartIconButtonType
-  showTotalItemsQty: MinicartTotalItemsType
+  itemCountMode: MinicartTotalItemsType
 }
 
 const MinicartIconButton: FC<MinicartIconButtonProps> = ({
   quantityDisplay,
-  showTotalItemsQty
+  itemCountMode,
 }) => {
   const { orderForm, loading }: OrderFormContext = useOrderForm()
   const handles = useCssHandles(CSS_HANDLES)
   const { open, openBehavior, openOnHoverProp } = useMinicartState()
   const dispatch = useMinicartDispatch()
-  
-  const totalItemsSum = (arr: Array<any>) =>
-    arr.reduce((sum: number, product: any) => sum + product.quantity, 0)
-  
-  const quantity = showTotalItemsQty === 'total' ? totalItemsSum(orderForm.items) : orderForm.items.length
+  const totalItemsSum = (arr: OrderFormItem[]) =>
+    arr.reduce((sum, product) => sum + product.quantity, 0)
+
+  const quantity =
+    itemCountMode === 'total'
+      ? totalItemsSum(orderForm.items)
+      : orderForm.items.length
   const itemQuantity = loading ? 0 : quantity
 
   const handleClick = () => {